+++ src/oci_attr/data_type.rs
@@ -363,7 +363,10 @@ unsafe impl DataType for MaxStringSize {
         match attr.get::<u8>()? {
             OCI_ATTR_MAXLEN_COMPAT_STANDARD => Ok(MaxStringSize::Standard),
             OCI_ATTR_MAXLEN_COMPAT_EXTENDED => Ok(MaxStringSize::Extended),
-            x => Err(Error::InternalError(format!("invalid MaxStringSize {}", x))),
+            x => Err(Error::internal_error(format!(
+                "invalid MaxStringSize {}",
+                x
+            ))),
         }
     }
 }