+++ src/sql_type/chrono.rs
@@ -32,7 +32,7 @@ use chrono::Duration;
 
 fn fixed_offset_from_sql(ts: &Timestamp) -> Result<FixedOffset> {
     FixedOffset::east_opt(ts.tz_offset())
-        .ok_or_else(|| Error::OutOfRange(format!("invalid time zone offset: {}", ts.tz_offset())))
+        .ok_or_else(|| Error::out_of_range(format!("invalid time zone offset: {}", ts.tz_offset())))
 }
 
 //
@@ -48,7 +48,7 @@ where
     date_from_sql(tz, ts)?
         .and_hms_nano_opt(ts.hour(), ts.minute(), ts.second(), ts.nanosecond())
         .ok_or_else(|| {
-            Error::OutOfRange(format!(
+            Error::out_of_range(format!(
                 "invalid year-month-day: {}-{}-{} {}:{}:{}.{:09}",
                 ts.year(),
                 ts.month(),
@@ -131,7 +131,7 @@ where
 {
     match tz.ymd_opt(ts.year(), ts.month(), ts.day()) {
         LocalResult::Single(date) => Ok(date),
-        _ => Err(Error::OutOfRange(format!(
+        _ => Err(Error::out_of_range(format!(
             "invalid month and/or day: {}-{}-{}",
             ts.year(),
             ts.month(),
@@ -198,7 +198,7 @@ fn naive_date_time_from_sql(ts: &Timestamp) -> Result<NaiveDateTime> {
     naive_date_from_sql(ts)?
         .and_hms_nano_opt(ts.hour(), ts.minute(), ts.second(), ts.nanosecond())
         .ok_or_else(|| {
-            Error::OutOfRange(format!(
+            Error::out_of_range(format!(
                 "invalid year-month-day: {}-{}-{} {}:{}:{}.{:09}",
                 ts.year(),
                 ts.month(),
@@ -249,7 +249,7 @@ impl ToSql for NaiveDateTime {
 
 fn naive_date_from_sql(ts: &Timestamp) -> Result<NaiveDate> {
     NaiveDate::from_ymd_opt(ts.year(), ts.month(), ts.day()).ok_or_else(|| {
-        Error::OutOfRange(format!(
+        Error::out_of_range(format!(
             "invalid year-month-day: {}-{}-{}",
             ts.year(),
             ts.month(),
@@ -289,7 +289,7 @@ impl ToSql for NaiveDate {
 impl FromSql for Duration {
     fn from_sql(val: &SqlValue) -> Result<Duration> {
         let err = |it: IntervalDS| {
-            Error::OutOfRange(format!(
+            Error::out_of_range(format!(
                 "unable to convert interval day to second {} to chrono::Duration",
                 it
             ))
@@ -336,7 +336,7 @@ impl ToSql for Duration {
         let minutes = secs / 60;
         let secs = secs % 60;
         if days.abs() >= 1000000000 {
-            return Err(Error::OutOfRange(format!("too large days: {}", self)));
+            return Err(Error::out_of_range(format!("too large days: {}", self)));
         }
         let it = IntervalDS::new(
             days as i32,